bihar-ai-chatbot/requirements.txt
streamlit>=1.28.0
transformers>=4.34.0
torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
gtts>=2.3.1
speechrecognition>=3.10.0
pydub>=0.25.1
googletrans>=4.0.0-rc1
streamlit-webrtc>=0.44.4
pyttsx3>=2.90